# mascotas-api/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Mascotas y Personas",
        "description": "Documentación de la API RESTful con autenticación JWT",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/personas": {
            "get": {
                "tags": [
                    "Personas"
                ],
                "summary": "Listar personas paginadas",
                "operationId": "19153d93fbbf375afea8d2a21f70831b",
                "responses": {
                    "200": {
                        "description": "Listado de personas"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Personas"
                ],
                "summary": "Crear nueva persona",
                "operationId": "b76ce2b01c2e821292fac2c4cddb74ee",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nombre",
                                    "email",
                                    "fecha_nacimiento"
                                ],
                                "properties": {
                                    "nombre": {
                                        "type": "string",
                                        "example": "Ana Pérez"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "ana@example.com"
                                    },
                                    "fecha_nacimiento": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "1990-01-01"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Persona creada exitosamente"
                    },
                    "422": {
                        "description": "Error de validación"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/personas/{id}": {
            "get": {
                "tags": [
                    "Personas"
                ],
                "summary": "Obtener detalles de una persona",
                "operationId": "9ae93a77b50e534e16e3a4aded57564d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la persona",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Persona encontrada"
                    },
                    "404": {
                        "description": "Persona no encontrada"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Personas"
                ],
                "summary": "Actualizar una persona",
                "operationId": "a4370a9476fe98f54efa656fbddd395b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la persona a actualizar",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nombre": {
                                        "type": "string",
                                        "example": "Ana María"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "ana.m@example.com"
                                    },
                                    "fecha_nacimiento": {
                                        "type": "string",
                                        "example": "1991-02-15"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Persona actualizada"
                    },
                    "404": {
                        "description": "Persona no encontrada"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Personas"
                ],
                "summary": "Eliminar una persona",
                "operationId": "65acb26d6a487dcc8cb74f199a297fe0",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Persona eliminada correctamente"
                    },
                    "404": {
                        "description": "Persona no encontrada"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/personas/{id}/mascotas": {
            "get": {
                "tags": [
                    "Personas"
                ],
                "summary": "Obtener persona con sus mascotas",
                "operationId": "e95bc34c0d45bf5067d3f72cb783d307",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la persona",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Persona con mascotas"
                    },
                    "404": {
                        "description": "Persona no encontrada"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}